Done with exercise 01
Logs have different levels which are used in different situations. The levels and their usages are listed below:

Verbose: Show all log messages (the default).
Debug: Show debug log messages that are useful during development only, as well as the message levels lower in this list.
Info: Show expected log messages for regular usage, as well as the message levels lower in this list.
Warn: Show possible issues that are not yet errors, as well as the message levels lower in this list.
Error: Show issues that have caused errors, as well as the message level lower in this list.
Assert: Show issues that the developer expects should never happen.